Fix EpochIOGroup count for idle CPUs

+++ src/EpochIOGroup.hpp
@@ -80,7 +80,7 @@ namespace geopm
             const PlatformTopo &m_topo;
             const ApplicationSampler &m_app;
             int m_num_cpu;
-            std::vector<int> m_per_cpu_count;
+            std::vector<double> m_per_cpu_count;
             std::map<int, std::set<int> > m_process_cpu_map;
             bool m_is_batch_read;
             bool m_is_initialized;
